Item set rarity implementation
Implemented item set rarity mentioned in dev cto chat. As a side effect, enemy location can be selected in battle mode and weapons are tied to the enemies you are fighting and not the mission, so in battle mode you can select west EU PMCs in a east EU terror attack mission and they will have the proper weapons.

Also m9 hand sprites.